+++ include/FmDecode.h
@@ -26,78 +26,11 @@
 #include "IfSimpleAgc.h"
 #include "MultipathFilter.h"
 #include "PhaseDiscriminator.h"
+#include "PilotPhaseLock.h"
 #include "SoftFM.h"
 
-/** Phase-locked loop for stereo pilot. */
-class PilotPhaseLock {
-public:
-  // Static constants.
-  // Expected pilot frequency (used for PPS events).
-  static constexpr int pilot_frequency = 19000;
-  // IF sampling rate.
-  static constexpr double sample_rate_if = 384000;
-  // Bandwidth (30Hz) relative to sample frequency.
-  static constexpr double bandwidth = 30 / sample_rate_if;
-  // Minimum pilot amplitude (lowered to prevent accidental unlocking)
-  static constexpr double minsignal = 0.001;
+// Complete decoder for FM broadcast signal.
 
-  /** Timestamp event produced once every 19000 pilot periods. */
-  struct PpsEvent {
-    std::uint64_t pps_index;
-    std::uint64_t sample_index;
-    double block_position;
-  };
-
-  /**
-   * Construct phase-locked loop.
-   *
-   * freq        :: 19 kHz center frequency relative to sample freq
-   *                (0.5 is Nyquist)
-   */
-  PilotPhaseLock(double freq);
-
-  /**
-   * Process samples and extract 19 kHz pilot tone.
-   * Generate phase-locked 38 kHz tone with unit amplitude.
-   * pilot_shift :: true to shift pilot phase by
-   *             :: using cos(2*x) instead of sin (2*x)
-   *             :: (for multipath distortion detection)
-   */
-  void process(const SampleVector &samples_in, SampleVector &samples_out,
-               bool pilot_shift);
-
-  /** Return true if the phase-locked loop is locked. */
-  bool locked() const { return m_lock_cnt >= m_lock_delay; }
-
-  /** Return detected amplitude of pilot signal. */
-  double get_pilot_level() const { return 2 * m_pilot_level; }
-
-  // Return calculated frequency error.
-  double get_freq_err() const { return m_freq_err; }
-
-  /** Return PPS events from the most recently processed block. */
-  std::vector<PpsEvent> get_pps_events() const { return m_pps_events; }
-
-  // Erase the first PPS event.
-  void erase_first_pps_event() { m_pps_events.erase(m_pps_events.begin()); }
-
-private:
-  Sample m_minfreq, m_maxfreq;
-  Sample m_freq, m_phase;
-  Sample m_pilot_level;
-  int m_lock_delay;
-  int m_lock_cnt;
-  int m_pilot_periods;
-  std::uint64_t m_pps_cnt;
-  std::uint64_t m_sample_cnt;
-  std::vector<PpsEvent> m_pps_events;
-  BiquadIirFilter m_biquad_phasor_i1, m_biquad_phasor_i2;
-  BiquadIirFilter m_biquad_phasor_q1, m_biquad_phasor_q2;
-  FirstOrderIirFilter m_first_phase_err;
-  Sample m_freq_err;
-};
-
-/** Complete decoder for FM broadcast signal. */
 class FmDecoder {
 public:
   // Static constants.
@@ -113,48 +46,47 @@ public:
   static constexpr double deemphasis_time_eu = 50; // Europe and Japan
   static constexpr double deemphasis_time_na = 75; // USA/Canada
 
-  /**
-   * Construct FM decoder.
-   *
-   * fmfilter_enable   :: True to enable IQSample filter frontend.
-   * fmfilter_coeff    :: IQSample filter coefficients.
-   * stereo            :: True to enable stereo decoding.
-   * deemphasis        :: Time constant of de-emphasis filter in microseconds
-   *                      (50 us for broadcast FM, 0 to disable de-emphasis).
-   * pilot_shift       :: True to shift pilot signal phase
-   *                   :: (use cos(2*x) instead of sin (2*x))
-   *                   :: (for multipath distortion detection)
-   * multipath_stages  :: Set >0 to enable multipath filter
-   *                   :: (LMS adaptive filter stage number)
-   */
+  //
+  // Construct FM decoder.
+  //
+  // fmfilter_enable   :: True to enable IQSample filter frontend.
+  // fmfilter_coeff    :: IQSample filter coefficients.
+  // stereo            :: True to enable stereo decoding.
+  // deemphasis        :: Time constant of de-emphasis filter in microseconds
+  //                      (50 us for broadcast FM, 0 to disable de-emphasis).
+  // pilot_shift       :: True to shift pilot signal phase
+  //                   :: (use cos(2*x) instead of sin (2*x))
+  //                   :: (for multipath distortion detection)
+  // multipath_stages  :: Set >0 to enable multipath filter
+  //                   :: (LMS adaptive filter stage number)
+  //
   FmDecoder(bool fmfilter_enable, IQSampleCoeff &fmfilter_coeff, bool stereo,
             double deemphasis, bool pilot_shift, unsigned int multipath_stages);
-  /**
-   * Process IQ samples and return audio samples.
-   *
-   * If the decoder is set in stereo mode, samples for left and right
-   * channels are interleaved in the output vector (even if no stereo
-   * signal is detected). If the decoder is set in mono mode, the output
-   * vector only contains samples for one channel.
-   */
+  //
+  // Process IQ samples and return audio samples.
+  //
+  // If the decoder is set in stereo mode, samples for left and right
+  // channels are interleaved in the output vector (even if no stereo
+  // signal is detected). If the decoder is set in mono mode, the output
+  // vector only contains samples for one channel.
   void process(const IQSampleVector &samples_in, SampleVector &audio);
 
-  /** Return true if a stereo signal is detected. */
+  // Return true if a stereo signal is detected.
   bool stereo_detected() const { return m_stereo_detected; }
 
-  /** Return actual frequency offset in Hz with respect to receiver LO. */
+  // Return actual frequency offset in Hz with respect to receiver LO.
   float get_tuning_offset() const { return m_baseband_mean * freq_dev; }
 
-  /** Return RMS baseband signal level (where nominal level is 0.707). */
+  // Return RMS baseband signal level (where nominal level is 0.707).
   float get_baseband_level() const { return m_baseband_level; }
 
-  /** Return amplitude of stereo pilot (nominal level is 0.1). */
+  // Return amplitude of stereo pilot (nominal level is 0.1).
   double get_pilot_level() const { return m_pilotpll.get_pilot_level(); }
 
   // Return RMS IF level.
   float get_if_rms() const { return m_if_rms; }
 
-  /** Return PPS events from the most recently processed block. */
+  // Return PPS events from the most recently processed block.
   std::vector<PilotPhaseLock::PpsEvent> get_pps_events() const {
     return m_pilotpll.get_pps_events();
   }
